Update signal usage in api package
